Rename paths and references from 'success' to 'return'
The API defines a single URL for the platform to return the user to regardless of whether the payment succeeded or failed. There was a mismatch between the API and this app which was using 'success' as the name of the return page.

This PR changes the names of things in this app to be consistent with the API.